Rename MultiMedium*Controller to MultiMedium*State
These are much more related to keeping track of the state of
a MultiMedium player rather than controlling it (even where there are
a few controls). This is also to avoid confusion with the
SingleMediumController, which is a proper controller on a playing media.
It also manages behaviour very specific to the Lunofono player while the
SingleMediumPlayer is more general purpose.

Signed-off-by: Leandro Lucarella <luca@llucax.com>