Create a SingleMediumPlayer
This player is directly extracted from MultiMediumTrackPlayer. It also
includes a big refactoring to separate all Player classes and States
into their own files, as the code is complex enough already to have all
of them mixed up. This also allows a simpler mapping with unit test
(which are also complex and was already splitting tests of different
classes in different files).

This will enable to follow the model more closely and being able to play
single media directly without having to wrap it in a MultiMedium.

Signed-off-by: Leandro Lucarella <luca@llucax.com>