Unify play/pause interface for MultiMedium(Track)State
MultiMediumState has no play/pause, playing is done automatically when
initializing. A more fine control is needed for playing Playlists, as
MultiMedium might need to be initialized but not played immediately, the same
as with SingleMedium in a MultiMediumTrack.

MultiMediumTrack provide playCurrent() and pauseCurrent(). The 'Current' suffix
provides little information as there is no playOther() or pauseOther(), so is
better to have a more uniform interface and call them play/pause as in all other
classes.

Signed-off-by: Leandro Lucarella <luca@llucax.com>