Make initialization interface for states more uniform
Initialization of different *State is done very differently. SingleMediumState
has a initialize() method that won't start playing, while initialize() for
MultiMediumState will. On top of that, MultiMediumTrackState calls the method
initializeAll(), and it doesn't start playing either. Then SingleMediumState
has a property to mark when initialization was done called isInitialized but
MultiMediumState calls it allInitialized and MultiMediumTrackState doesn't have
any.

This commit makes the initialization interface more uniform by naming all
methods initialize() and only start playing the medium if it's explicitly
requested via the new optional argument startPlaying. The property to mark when
initialization was done is renamed or added to be isInitialized for all too.

Finally, the notifyListeners() method is only called when the state really
changes for the State object, and not when calls are forwarded to internal
state tracking object (like the SingleMediumController) to avoid unnecessary
widget updates.

Signed-off-by: Leandro Lucarella <luca@llucax.com>