Don't propagate SingleMediumControllers
Passing around SingleMediumControllers everywhere just to be able to test stuff
is not really practical, and in dart it makes a lot of sense to just define
global instances that can be overridden for testing because there are no
concurrency issues, so instead we just do that. ControllerRegistry.instance
becomes writable so it can be overridden by tests and the library always use
the global instance instead of offering a way to override the registry in the
constructor.

This also means now SingleMediumState is in charge of creating the controller,
which has also the nice side-effect that SingleMediumState is self-contained and
a SingleMedium without a registered controller is just one more kind of error
state it can be in, so there is no need to handle that kind of errors in
MultiMediumTrackState.

The SingleMediumPlayer is also greatly simplified as it doesn't need to deal
with media without a registered controller.

SingleMediumStateFactory is also completely wiped out in favour of this
approach. The overrideable MultiMediumTrackState.createSingleMediumState
callback is added to create SingleMediumStates for testing.

Finally, not very related, but the widgetKey and build members of
SingleMediumState are removed in favour of just using the controller, as they
are just forwarding to it, there are not performing any state update.

Signed-off-by: Leandro Lucarella <luca@llucax.com>