Replace video_player with audioplayers (partial)
The `video_player` plugin doesn't officially support playing audio. In
the long term, it should be replace with the `just_audio` plugin, but
this plugin uses ExoPlayer as the Android backed, which has a bug when
playing short sounds (#15). `video_player` uses ExoPlayer too and we
need to play short songs so we need a temporary solution.

The `audioplayers` plugin uses MediaPlayer as the Android backend which
works with short sounds, but it can't play assets in web, which we also
need.

Because of this we temporarily replace the `video_player` plugin with
`audioplayers` but only for the non-web platforms, so playing assets in
web still works.

Fixes #15.

Signed-off-by: Leandro Lucarella <luca@llucax.com>