ci: Be more permissive with coverage
For now we managed to have 100% code coverage, but this is not always
feassible (or at least practical). We will soon introduce a new audio
backend temporarily, and it requires a lot of effort to write tests for
it, but this change is only temporary so it doesn't make much sense to
go through all that effort.

Because of this we are relaxing the coverage requirements for Codecov
(now it succeeds PRs even if they reduce coverage by up to 2.5% and the
PR is only 60% covered) and the tool/ci script (requires coverage higher
than 97%).

Signed-off-by: Leandro Lucarella <luca@llucax.com>