codecov: Add patch configuration
On 92b107b677df9e25415141d793c946daa026aea8 we tried to be more
permissive with changes introduced in PRs, but we configured only the
options for the whole project, not just patches.

This commit sets the minimum global project coverage to 95% and patches
with a minimum 60% coverage are still accepted.