Remove flutter_grid_button dependency
Now the GridMenuPlayer doesn't use the flutter_grid_button anymore and
it implements its own GridMenuWidget to render the rows and columns.
Thanks to this we have full control on how the grid is presented, so the
divider lines are also removed and the buttons can show themselves more
autonomously as regular flutter widgets.

This fixes #10 and #11.

Signed-off-by: Leandro Lucarella <luca@llucax.com>