ci: Stop using kitek/dartanalyzer-annotations-action
The dartanalyzer has a bug when excluding files that have been around
for a long time now, and since it is now deprecated in flavour of `dart
analyze` it is very unlikely to be fixed any time soon.

Since this action uses dartanalyzer and we need to exclude mocks, we
can't use it anymore.

https://github.com/dart-lang/sdk/issues/25551